+++ src/pages/topics/[slug].tsx
@@ -6,7 +6,7 @@ import theme from '@mdb/flora/theme';
 import { GridLayout } from '@mdb/flora';
 
 import Hero from '../../components/hero';
-import Search from '../../components/search';
+import Search from '../../components/search/search';
 import { CTA } from '../../interfaces/components/hero';
 
 interface TopicProps {
